[PPO] Copy code directly from baselines.
Made enough PIRL-specific changes to it that it is more appropriate to
have it in our codebase rather than maintaining a fork of baselines.

New change is to use the reward seen by PPO for validation, rather than
the reward reported by bench.Monitor (that will be before wrapping in
the case of reoptimization).